Work in progress

+++ chapter1/string_vector/string_vector_large.h
@@ -47,7 +47,7 @@ int string_vector_large_add(StringVectorLarge string_vector_large,
  * Gets the string at position
  */
 char *string_vector_large_get(StringVectorLarge string_vector_large,
-		unsigned position);
+		unsigned long position);
 
 /**
  * Swaps the elements in two positions
@@ -56,7 +56,7 @@ int string_vector_large_swap(StringVectorLarge string_vector_large,
 		unsigned position1, unsigned position2);
 
 /**
- * Frees all memory allocated to the StringVector
+ * Frees all memory allocated
  */
 void string_vector_large_free(StringVectorLarge string_vector_large);
